+++ test/unit/lib/networkHandlerTests.js
@@ -1322,6 +1322,41 @@ describe('networkHandler', () => {
                 });
         });
 
+        it('should send out 1 enable value to the sys/db/...routing endpoint if a declaration includes RoutingAccessList', () => {
+            const declaration = {
+                Common: {
+                    RoutingAccessList: {
+                        RoutingAccessList1: {
+                            name: 'RoutingAccessList1',
+                            entries: [
+                                {
+                                    name: 10
+                                }
+                            ]
+                        },
+                        RoutingAccessList2: {
+                            name: 'RoutingAccessList2',
+                            entries: [
+                                {
+                                    name: 20
+                                },
+                                {
+                                    name: 30
+                                }
+                            ]
+                        }
+                    }
+                }
+            };
+
+            const networkHandler = new NetworkHandler(declaration, bigIpMock);
+            return networkHandler.process()
+                .then(() => {
+                    const data = dataSent['/tm/sys/db/tmrouted.tmos.routing'];
+                    assert.deepStrictEqual(data, [{ value: 'enable' }]);
+                });
+        });
+
         it('should send out 1 enable value to the sys/db/...routing endpoint if a declaration includes RoutingPrefixList', () => {
             const declaration = {
                 Common: {
@@ -1444,7 +1479,7 @@ describe('networkHandler', () => {
                 });
         });
 
-        it('should send out 1 enable value to the sys/db/...routing endpoint if a declaration includes RoutingAsPath, RoutingPrefixList, RouteMap, and RoutingBGP', () => {
+        it('should send out 1 enable value to the sys/db/...routing endpoint if a declaration includes RoutingAsPath, RoutingAccessList, RoutingPrefixList, RouteMap, and RoutingBGP', () => {
             const state = {
                 currentConfig: {
                     Common: {}
@@ -1475,6 +1510,12 @@ describe('networkHandler', () => {
                                     regex: '^$'
                                 }
                             ]
+                        }
+                    },
+                    RoutingAccessList: {
+                        RoutingAccessList1: {
+                            name: 20,
+                            entries: []
                         }
                     },
                     RoutingPrefixList: {
@@ -1651,6 +1692,177 @@ describe('networkHandler', () => {
                     assert.deepStrictEqual(data[0], {
                         name: 'RoutingAsPath1',
                         partition: 'Common',
+                        entries: {}
+                    });
+                });
+        });
+    });
+
+    describe('RoutingAccessList', () => {
+        it('should handle a fully specified Routing Access List', () => {
+            const declaration = {
+                Common: {
+                    RoutingAccessList: {
+                        RoutingAccessList1: {
+                            name: 'RoutingAccessList1',
+                            description: 'my description 1',
+                            entries: [
+                                {
+                                    name: 10,
+                                    action: 'permit',
+                                    destination: '10.2.2.0/24',
+                                    exactMatch: 'disabled',
+                                    source: '10.3.3.0/24'
+                                },
+                                {
+                                    name: 20,
+                                    action: 'deny',
+                                    destination: '10.4.4.1/32',
+                                    exactMatch: 'disabled',
+                                    source: '10.4.4.2/32'
+                                }
+                            ]
+                        },
+                        RoutingAccessList2: {
+                            name: 'RoutingAccessList2',
+                            description: 'my description 2',
+                            entries: [
+                                {
+                                    name: 30,
+                                    action: 'permit',
+                                    destination: '0.0.0.0/0',
+                                    exactMatch: 'enabled',
+                                    source: '10.5.5.1'
+                                }
+                            ]
+                        },
+                        RoutingAccessList3: {
+                            name: 'RoutingAccessList3',
+                            description: 'my description 3',
+                            entries: [
+                                {
+                                    name: 40,
+                                    action: 'permit',
+                                    destination: '1111:1111::/64',
+                                    exactMatch: 'disabled',
+                                    source: '1111:3333::/64'
+                                },
+                                {
+                                    name: 50,
+                                    action: 'deny',
+                                    destination: '1111:2222::/128',
+                                    exactMatch: 'disabled',
+                                    source: '1111:3333::/128'
+                                }
+                            ]
+                        },
+                        RoutingAccessList4: {
+                            name: 'RoutingAccessList4',
+                            entries: [
+                                {
+                                    name: 60,
+                                    action: 'permit',
+                                    destination: '::',
+                                    exactMatch: 'enabled',
+                                    source: '1111:3333::/64'
+                                }
+                            ]
+                        }
+                    }
+                }
+            };
+
+            const networkHandler = new NetworkHandler(declaration, bigIpMock);
+            return networkHandler.process()
+                .then(() => {
+                    const data = dataSent[PATHS.RoutingAccessList];
+                    assert.deepStrictEqual(data[0], {
+                        name: 'RoutingAccessList1',
+                        partition: 'Common',
+                        description: 'my description 1',
+                        entries: {
+                            10: {
+                                action: 'permit',
+                                destination: '10.2.2.0/24',
+                                exactMatch: 'disabled',
+                                source: '10.3.3.0/24'
+                            },
+                            20: {
+                                action: 'deny',
+                                destination: '10.4.4.1/32',
+                                exactMatch: 'disabled',
+                                source: '10.4.4.2/32'
+                            }
+                        }
+                    });
+                    assert.deepStrictEqual(data[1], {
+                        name: 'RoutingAccessList2',
+                        partition: 'Common',
+                        description: 'my description 2',
+                        entries: {
+                            30: {
+                                action: 'permit',
+                                destination: '0.0.0.0/0',
+                                exactMatch: 'enabled',
+                                source: '10.5.5.1'
+                            }
+                        }
+                    });
+                    assert.deepStrictEqual(data[2], {
+                        name: 'RoutingAccessList3',
+                        partition: 'Common',
+                        description: 'my description 3',
+                        entries: {
+                            40: {
+                                action: 'permit',
+                                destination: '1111:1111::/64',
+                                exactMatch: 'disabled',
+                                source: '1111:3333::/64'
+                            },
+                            50: {
+                                action: 'deny',
+                                destination: '1111:2222::/128',
+                                exactMatch: 'disabled',
+                                source: '1111:3333::/128'
+                            }
+                        }
+                    });
+                    assert.deepStrictEqual(data[3], {
+                        name: 'RoutingAccessList4',
+                        partition: 'Common',
+                        description: 'none',
+                        entries: {
+                            60: {
+                                action: 'permit',
+                                destination: '::',
+                                exactMatch: 'enabled',
+                                source: '1111:3333::/64'
+                            }
+                        }
+                    });
+                });
+        });
+
+        it('should handle empty entries property', () => {
+            const declaration = {
+                Common: {
+                    RoutingAccessList: {
+                        RoutingAccessList1: {
+                            name: 'RoutingAccessList1',
+                            entries: []
+                        }
+                    }
+                }
+            };
+
+            const networkHandler = new NetworkHandler(declaration, bigIpMock);
+            return networkHandler.process()
+                .then(() => {
+                    const data = dataSent[PATHS.RoutingAccessList];
+                    assert.deepStrictEqual(data[0], {
+                        name: 'RoutingAccessList1',
+                        partition: 'Common',
+                        description: 'none',
                         entries: {}
                     });
                 });