+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes to this project are documented in this file. More detail and links can b
 - AUTOTOOL-2675: ([GitHub Issue 241](https://github.com/F5Networks/f5-declarative-onboarding/issues/241)): Add chargeBackTag to License class
 - AUTOTOOL-2676: ([GitHub Issue 218](https://github.com/F5Networks/f5-declarative-onboarding/issues/218)): Support for configuring management-ip-rules with ManagementIpFirewall class
 - AUTOTOOL-705: SnmpTrapDestination default values for version, port, and network
+- AUTOTOOL-2259: RoutingAccessList (net routing access-list)
 
 ### Fixed
 - AUTOTOOL-2692: FirewallPolicy incorrectly allows VLANs to be included in the destination schema object